+++ sdk/core/typespec_client_core/examples/core_binary_data_request.rs
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-mod common;
-
-use common::fs::FileStreamBuilder;
 use tokio::fs;
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::fmt::format::FmtSpan;
-use typespec_client_core::http::RequestContent;
+use typespec_client_core::{
+    http::{Body, RequestContent},
+    stream::{tokio::FileReader, FileStream},
+};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -23,12 +23,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Asynchronously stream the file with the service client request.
     let file = fs::File::open(file!()).await?;
-    let file = FileStreamBuilder::new(file)
-        // Simulate a slow, chunky request.
-        .buffer_size(512usize)
-        .build()
-        .await?;
-    client::put_binary_data(file.into()).await?;
+    let reader = FileReader::from(file);
+    // Simulate a slow, chunky request.
+    let stream = FileStream::from(reader).with_buffer_size(512);
+    let body: Body = Body::SeekableStream(Box::new(stream));
+    client::put_binary_data(body.into()).await?;
 
     Ok(())
 }